feat: remove embedded binaries
