+++ DevantlerTech.TalosctlCLI.slnx
@@ -1,0 +1,8 @@
+<Solution>
+  <Folder Name="/src/">
+    <Project Path="src/DevantlerTech.TalosctlCLI/DevantlerTech.TalosctlCLI.csproj" />
+  </Folder>
+  <Folder Name="/tests/">
+    <Project Path="tests/DevantlerTech.TalosctlCLI.Tests/DevantlerTech.TalosctlCLI.Tests.csproj" />
+  </Folder>
+</Solution>
