--- src/Devantler.TalosctlCLI/runtimes/linux-arm64/native/talosctl-linux-arm64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b22e1b99f984963d81be10729d5895e2f0b2cd04e07be3096537595941c1ffb9
-size 89850040
