--- src/Devantler.TalosctlCLI/runtimes/linux-x64/native/talosctl-linux-x64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c5d89a262fd00c0d6c01df573c20b70b62b053f708dcafebb319daaa485a7e1b
-size 93933752
