--- src/Devantler.TalosctlCLI/runtimes/osx-arm64/native/talosctl-osx-arm64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:556d62984652b8017c120dc28d7dff530e255970829e06721b54becfa04bca86
-size 89624194
