--- src/Devantler.TalosctlCLI/runtimes/osx-x64/native/talosctl-osx-x64
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:868cadab8df3e506852da6bb62d67046e77039b785873f40a470b9b136c4c2a7
-size 92935680
