--- src/Devantler.TalosctlCLI/runtimes/win-arm64/native/talosctl-win-arm64.exe
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d54eac3224173c921670d8a08262c86e7d993ef4e11b9023a7005a854718b35a
-size 87582720
