--- src/Devantler.TalosctlCLI/runtimes/win-x64/native/talosctl-win-x64.exe
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d60a3d9985047bd08c3fd022b19c7ee9ad82a614a95caabb207c11efe6039fe9
-size 92771840
